[MSHADE-391] Add integration test
Test src/it/projects/MSHADE-391_noRelocationKeepOriginalClasses shades
two dependencies, relocating one of them. For the non-relocated one, it
compares pre-calculated MD5 checksums to the actual ones in the uber
JAR, making sure that they are identical. The test will fail in Maven
Shade 3.2.4 or older, but pass with the changes made in MSHADE-391.