[MSHADE-391] Refactor previous commit @50e101fe
There was some suboptimal naming: We had class names
  - Relocator,
  - SimpleRelocator,
  - Relocators,
  - DefaultRelocators.

That was a bit too much of '*Relocator(s)', and the plural was not nice
either. We are actually dealing with one mapper (interface), the default
implementation of which happens to own a list of relocators. We now have
  - Relocator,
  - SimpleRelocator,
  - PackageMapper,
  - DefaultPackageMapper.

I also changed PackageMapper.map to have two boolean parameters now,
even though the first one is always true in current use cases. But I
wanted the method signature to more clearly describe what the boolean
parameters do. Before there was just 'boolean clazz', which does not
communicate anything to the caller. Now we have the method signature

String map(String entityName, boolean mapPaths, boolean mapPackages)

which is better in this respect. Also, 'entityName' is a clearer
than just 'name'. Despite the interface being private, I also added
Javadoc to this method, because it is a crucial one for understanding
how class relocation and entity name mapping work.

Test case DefaultShaderTest.testNoopWhenNotRelocated was refactored to
be a bit more readable and also to explicitly test that the byte code
for a relocated class has actually changed and not only its path within
the uber JAR has been adjusted. I also renamed the helper method
'areEquals' to 'areEqual'.