[MSHADE-396] Improve SourceContent Shading
Improve search & replace heuristics without destroying previously
correct replacements in my test project (AspectJ). This solution is
still bound to fail in some situations, simply because it is just a
heuristic approach and not a full Java parser correctly recognising
package names in all possible situations in Java source code. As for
matching within Java string constants, this is next to impossible to get
100% right.

But the source shading feature is not meant as a source code
generator anyway, merely as a tool creating reasonably plausible source
code when navigating to a relocated library class from an IDE, hopefully
displaying source code which makes 95% sense - no more, no less.