[MSHADE-323] compatible with the flatten plugin
The runtime injected pom will be set via `MavenProject::setFile` or
`MavenProject::setPomFile`, including this plugin and the flatten plugin.
So to chain these kind of plugins, when generating the d-r-p pom, this
pr first try to parse a model directly from `MavenProject::getFile` and
fallback to a clone of `MavenProject::getOriginalModel` if not set the
pom file.